+++ app/components/Views/confirmations/components/info/switch-account-type/switch-account-type.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { View } from 'react-native';
 
-import { ConfirmationInfoComponentIDs } from '../../../constants/test';
+import { ConfirmationInfoComponentIDs } from '../../../constants/info-ids';
 import { useConfirmationMetricEvents } from '../../../hooks/metrics/useConfirmationMetricEvents';
 import AccountNetworkInfoRow from '../../rows/account-network-info-row';
 import AdvancedDetailsRow from '../../rows/transactions/advanced-details-row/advanced-details-row';
