+++ app/components/Views/confirmations/components/info/typed-sign-v3v4/typed-sign-v3v4.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View } from 'react-native';
 
-import { ConfirmationInfoComponentIDs } from '../../../constants/test';
+import { ConfirmationInfoComponentIDs } from '../../../constants/info-ids';
 import AccountNetworkInfoRow from '../../rows/account-network-info-row';
 import { InfoSectionOriginAndDetails } from './info-section-origin-and-details';
 import Message from './message';
