+++ app/components/Views/confirmations/external/staking/info/staking-deposit/staking-deposit.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { View } from 'react-native';
 import { strings } from '../../../../../../../../locales/i18n';
 import { EVENT_PROVIDERS } from '../../../../../../UI/Stake/constants/events';
-import { ConfirmationInfoComponentIDs } from '../../../../constants/test';
+import { ConfirmationInfoComponentIDs } from '../../../../constants/info-ids';
 import useClearConfirmationOnBackSwipe from '../../../../hooks/ui/useClearConfirmationOnBackSwipe';
 import { useConfirmationMetricEvents } from '../../../../hooks/metrics/useConfirmationMetricEvents';
 import useNavbar from '../../../../hooks/ui/useNavbar';
